Implemented SR incentive based on reseed requests.
Whenever a user fails to download a file because of lack of
peers, the file is added to Tracker's list of requested files.

Any seeder fulfilling a reseed request gets multiplied bonus per upload.
Thus, a seeder now always checks if it can fulfill a request and if it can,
it does so.